Remove unused statement resource thread pools from workers
The statement resource is only bound on the coordinator.
